Refactor pyproject package collection helpers
