+++ pkg/lockfile/python/parse-pyproject-toml.go
@@ -1,13 +1,9 @@
 package python
 
 import (
-	"cmp"
 	"fmt"
 	"io"
-	"log"
-	"maps"
 	"path/filepath"
-	"slices"
 	"strings"
 
 	"github.com/BurntSushi/toml"
@@ -226,134 +222,6 @@ func (e PyProjectTOMLExtractor) Extract(f lockfile.DepFile, context lockfile.Sca
 	}
 
 	return sortedPyprojectPackages(collector.packages), nil
-}
-
-type pyprojectPackageCollector struct {
-	packages       map[string]lockfile.PackageDetails
-	lines          []string
-	path           string
-	packageManager models.PackageManager
-}
-
-func (c *pyprojectPackageCollector) addDependency(dependency pep508Dependency, groups []string, isPoetry bool) {
-	if (dependency.Version == "") == (dependency.VersionRange == "") {
-		log.Printf(
-			"Skipping pyproject dependency %q from %s: expected exactly one of version or version range, got version=%q versionRange=%q\n",
-			dependency.Name,
-			c.path,
-			dependency.Version,
-			dependency.VersionRange,
-		)
-
-		return
-	}
-
-	block, nameLocation, versionLocation := extractPositions(c.lines, c.path, dependency.RawName, versionOrRange(dependency.Version, dependency.VersionRange), isPoetry)
-	c.addOrMergePackageGroups(lockfile.PackageDetails{
-		Name:                         dependency.Name,
-		Version:                      dependency.Version,
-		VersionRange:                 dependency.VersionRange,
-		PackageManager:               c.packageManager,
-		Ecosystem:                    models.EcosystemPyPI,
-		IsDirect:                     true,
-		RequiresTransitiveEnrichment: true,
-		DepGroups:                    groups,
-		BlockLocation:                block,
-		NameLocation:                 nameLocation,
-		VersionLocation:              versionLocation,
-		LocationRole:                 models.LocationRoleManifest,
-	})
-}
-
-// addOrMergePackageGroups adds a package to the map, or if it already exists (same name+version/range),
-// merges the new dep groups into the existing entry rather than dropping the duplicate.
-func (c *pyprojectPackageCollector) addOrMergePackageGroups(pkg lockfile.PackageDetails) {
-	key := pkg.Name + "@" + pkg.Version + "|" + pkg.VersionRange
-	if existing, exists := c.packages[key]; exists {
-		for _, g := range pkg.DepGroups {
-			if !slices.Contains(existing.DepGroups, g) {
-				existing.DepGroups = append(existing.DepGroups, g)
-			}
-		}
-		c.packages[key] = existing
-
-		return
-	}
-	c.logVersionRangeConflict(pkg)
-	c.packages[key] = pkg
-}
-
-func (c *pyprojectPackageCollector) logVersionRangeConflict(pkg lockfile.PackageDetails) {
-	if pkg.VersionRange == "" {
-		return
-	}
-
-	for _, existing := range c.packages {
-		if existing.Name != pkg.Name || existing.VersionRange == "" || existing.VersionRange == pkg.VersionRange {
-			continue
-		}
-
-		log.Printf(
-			"Multiple pyproject version ranges for dependency %q from %s collapse to the same unversioned PURL; CycloneDX output will keep the earliest source declaration. Saw ranges %q and %q\n",
-			pkg.Name,
-			c.path,
-			existing.VersionRange,
-			pkg.VersionRange,
-		)
-
-		return
-	}
-}
-
-func sortedPyprojectPackages(packages map[string]lockfile.PackageDetails) []lockfile.PackageDetails {
-	result := slices.Collect(maps.Values(packages))
-	slices.SortFunc(result, comparePyprojectPackageDetails)
-
-	return result
-}
-
-func comparePyprojectPackageDetails(a, b lockfile.PackageDetails) int {
-	if c := strings.Compare(a.BlockLocation.Filename, b.BlockLocation.Filename); c != 0 {
-		return c
-	}
-	if c := cmp.Compare(sourceLine(a.BlockLocation), sourceLine(b.BlockLocation)); c != 0 {
-		return c
-	}
-	if c := cmp.Compare(sourceColumn(a.BlockLocation), sourceColumn(b.BlockLocation)); c != 0 {
-		return c
-	}
-	if c := strings.Compare(a.Name, b.Name); c != 0 {
-		return c
-	}
-	if c := strings.Compare(a.Version, b.Version); c != 0 {
-		return c
-	}
-
-	return strings.Compare(a.VersionRange, b.VersionRange)
-}
-
-func sourceLine(location models.FilePosition) int {
-	if location.Line.Start == 0 {
-		return 1 << 30
-	}
-
-	return location.Line.Start
-}
-
-func sourceColumn(location models.FilePosition) int {
-	if location.Column.Start == 0 {
-		return 1 << 30
-	}
-
-	return location.Column.Start
-}
-
-func versionOrRange(version, versionRange string) string {
-	if version != "" {
-		return version
-	}
-
-	return versionRange
 }
 
 type pep508Dependency struct {
